+++ src/Network.php
@@ -4,6 +4,7 @@ namespace duncan3dc\Sonos;
 
 use duncan3dc\DomParser\XmlParser;
 use duncan3dc\Sonos\Devices\Discovery;
+use duncan3dc\Sonos\Devices\Speakers;
 use duncan3dc\Sonos\Exceptions\NotFoundException;
 use duncan3dc\Sonos\Exceptions\UnknownGroupException;
 use duncan3dc\Sonos\Interfaces\AlarmInterface;
@@ -51,7 +52,7 @@ final class Network implements NetworkInterface, LoggerAwareInterface
     public function __construct(?CollectionInterface $collection = null)
     {
         if ($collection === null) {
-            $collection = new Discovery();
+            $collection = new Speakers(new Discovery());
         }
         $this->collection = $collection;
     }
@@ -89,10 +90,6 @@ final class Network implements NetworkInterface, LoggerAwareInterface
 
         $this->speakers = [];
         foreach ($devices as $device) {
-            if (!$device->isSpeaker()) {
-                continue;
-            }
-
             $speaker = new Speaker($device);
 
             try {