+++ tests/SpeakerTest.php
@@ -234,7 +234,6 @@ class SpeakerTest extends AbstractMockCase
     {
         $device = $this->getDevice();
 
-        $device->shouldReceive("isSpeaker")->once()->andReturn(true);
         $device->shouldReceive("soap")->with("ZoneGroupTopology", "GetZoneGroupAttributes", [])->andReturn([
             "CurrentZoneGroupID" => "RINCON_5CAAFD472E1C01400:195",
             "CurrentZonePlayerUUIDsInGroup" => "RINCON_5CAAFD472E1C01400",
@@ -248,7 +247,6 @@ class SpeakerTest extends AbstractMockCase
     {
         $device = $this->getDevice();
 
-        $device->shouldReceive("isSpeaker")->once()->andReturn(true);
         $device->shouldReceive("soap")->with("ZoneGroupTopology", "GetZoneGroupAttributes", [])->andReturn([
             "CurrentZoneGroupID" => "RINCON_5CAAFD472E1C01400:195",
             "CurrentZonePlayerUUIDsInGroup" => "RINCON_5CAAF0000A1A01400",
@@ -262,7 +260,6 @@ class SpeakerTest extends AbstractMockCase
     {
         $device = $this->getDevice();
 
-        $device->shouldReceive("isSpeaker")->once()->andReturn(true);
         $device->shouldReceive("soap")->with("ZoneGroupTopology", "GetZoneGroupAttributes", [])->andReturn([
             "CurrentZoneGroupID" => "RINCON_5CAAFD472E1C01400:195",
             "CurrentZonePlayerUUIDsInGroup" => "RINCON_5CAAFD472E1C01400,RINCON_5CAAF0000A1A01400",
@@ -276,7 +273,6 @@ class SpeakerTest extends AbstractMockCase
     {
         $device = $this->getDevice();
 
-        $device->shouldReceive("isSpeaker")->once()->andReturn(true);
         $device->shouldReceive("soap")->with("ZoneGroupTopology", "GetZoneGroupAttributes", [])->andReturn([
             "CurrentZoneGroupID" => "RINCON_5CAAF0000A1A01400:195",
             "CurrentZonePlayerUUIDsInGroup" => "RINCON_5CAAFD472E1C01400,RINCON_5CAAF0000A1A01400",
@@ -292,7 +288,6 @@ class SpeakerTest extends AbstractMockCase
     {
         $device = $this->getDevice();
 
-        $device->shouldReceive("isSpeaker")->once()->andReturn(true);
         $device->shouldReceive("soap")->with("ZoneGroupTopology", "GetZoneGroupAttributes", [])->andReturn([
             "CurrentZoneGroupID" => "",
             "CurrentZonePlayerUUIDsInGroup" => "RINCON_5CAAFD472E1C01400",